apps/api/requirements.txt
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
PyYAML==6.0.2
boto3>=1.34.0

PyJWT>=2.8.0
